alsa-lib: add patch to support a rootfs mounted as ext2 on a loop device
In the case of our setup for the tuna device we're mounting the rootfs as ext2 on top of a
loop device which will let scandir report DT_UNKNOWN as type for every file and directory.
To work around this the implementation has to handle DT_UNKNOWN propably and determine the
correct file type with stat when DT_UNKNOWN occurs.

The implementation fixes the bug but seems to be unsuitable for upstream submission as it
adds unneeded redundant code. For a accepted upstream patch the whole scandir logic has to
be reworked to handle the case of DT_UNKNOWN propably.

Signed-off-by: Simon Busch <morphis@gravedo.de>