alsa-lib: import UCM support patches from ubuntu touch
This imports the machine UCM config as part of the alsa-lib package as it's easier to main
it this way to stay in sync with ubuntu touch. Having the UCM configuration for a
different machine than the one which is running doesn't have any impact. The audio stack
will automatically choose the correct configuration.

All three patches are taken from https://code.launchpad.net/~ubuntu-audio-dev/alsa-lib/ubuntu
at revision 173.

Signed-off-by: Simon Busch <morphis@gravedo.de>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>