qtwebkit: bump SRCREV and update patches
We're switching back to qtwebkit as upstream webkit doesn't offer us any new features we
don't get here as well and qtwebkit offers us a stable and tested and not always in
development component.

This enables the JIT again but disables the low-level interperter as it's causing problems
which lead to segfaults at runtime.

Furthermore we're now implementing the PalmServiceBridge inside WebKit instead of having
it's implementation as part of the WebAppMgr.

Signed-off-by: Simon Busch <morphis@gravedo.de>