nodejs: Fix segfault caused by using incompatible mksnapshot
:Release Notes:
Restrict fix for qemux86-64 only for cases where TARGET_ARCH and
BUILD_ARCH are the same.

:Detailed Notes:
When 64bit mksnapshot is used when building 32bit nodejs for qemux86
it is failing with segfault in runtime. Use mksnapshot from
nodejs-native only when it's the same architecture.

:Testing Performed:
Tested with qemux86 and qemux86-64 built on 64bit builder. Will try
to create image for both also on 32bit builder.

:QA Notes:
Nodejs services should work correctly again.

:Issues Addressed:
[GF-42389] Node services appear to be broken in build 352

Change-Id: I9b936d9f1aa3b4b443b969e9de141e33cb25d373
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>