fixed issue 559
See: http://issues.webos-ports.org/issues/559

This was caused by a corrupt .json files in /etc/palm/tempdb. Strange thing is that this file was not changed for ages... probably it was not read until recently, because it is in a subfolder? Anyhow, with this change the luna-send calls and reboot both work to configure db kinds and permissions.