pulseaudio: disable pthread priority inheritance
Programs using pulseaudio audio were crashing cause of pthread mutex problems. The
actual problem seems to be inside the pthread implementation of eglibc but we can
workaround this with disabling pthread priority inheritance for pulseaudio.

This problem seems to be specific for ARM as it doesn't appear in the x86/x64 based
emulator.

Signed-off-by: Simon Busch <morphis@gravedo.de>