initramfs-boot-android: fixup old rootfs directory after update
With the recent change of the distro name used in the initramfs we need to take care about
the update case too. If someone still has a rootfs directory named webos then we need to
move it to the correct new one as otherwise the init script is not able to find it. Same
applies for the data storage location.

Signed-off-by: Simon Busch <morphis@gravedo.de>