webos-keyboard: bump SRCREV
c681ccd Fix missing qmldir in install section of qml.pro
4094207 Tests: show what key has been sent to the system
9e0d0d8 Add qmldir to the keys .pro
95d02c1 Tests: make the test project also work for email and url layouts
bd98bc8 Tests: add a zone above the keyboard to interact with the settings of the latter. It also helps checking if the computed visibleRect of the vkb is correct.
63ea9bb Tests: improve test integration with original source code
e7d9b09 Tests: added a keyboard-test qmlproject file to be able to visualize a vkb on desktop. This might help people trying to improve the design of the vkb.

Signed-off-by: Simon Busch <morphis@gravedo.de>