luneos-image: add read-only support
This enables read-only support for the luneos-dev-image by default for all machines except
the emulator as we mainly don't need it there. Read-only support is necessary to prevent
any modifications to the root filesystem which could prevent us from applying any update
on top of the rootfs. If a developer needs read-write access he can simply remounting the
root filesystem with

$ mount -o remount,rw /

if he has root permissions.

Signed-off-by: Simon Busch <morphis@gravedo.de>