ACG: Initial work for migration
webOS OSE introduced a new version of the luna-service2 component with enhanced
security model called ACG (Access Control Groups).

ACG (Access Control Groups) is a new security model that provides
fine-grained access control of security permissions for services running
on Luna Bus. This allows developers to define separate security groups
with different set of methods based on functionality or security level.

Each service will specify it's API calls, the group(s) to which these
API calls belong to and the service names it uses on the Luna Bus. It
will also specify which permissions is requires to function properly.

This means developers need to set appropriate
ACG information when configuring apps or services, as follows:

For web apps, QML apps, and native apps
=======================================

- If methods of the external services are used, specify the group
information of the external methods in the requiredPermissions field of
appinfo.json.

For external JS services
========================

- If methods of the external services are used, add the group information
of the external methods to the requiredPermissions field in appinfo.json
of the web app used for packaging the JS service.

For built-in JS services and native services
============================================

- If methods of the external services are used, specify the group
information of the external methods in the client permission file.
- Specify relevant group and permission information in the role file and
API permission file.

Signed-off-by: Herman van Hazendonk <github.com@herrie.org>