Cleanup webOS-ports/webOS-OSE branch names by adding a separate .bbclass
Since for many repos we needed to make significant changes to them in order to integrate the webOS OSE changes or we needed to make a new branch since it doesn't share a commit history with the original Open webOS fork, we have created webOS-ports/webOS-OSE branches for these.

To more easily facilitate this, now added a bbclass called webos_ports_ose_repo which can be inherited so no need to specify the WEBOS_GIT_PARAM_BRANCH anymore in each of the recipes.

Signed-off-by: Herman van Hazendonk <github.com@herrie.org>