sensorfw: bring back custom setup for Halium devices
This bbappend was split from the main bb recipe in order to better
separate the setup specific to meta-smartphone devices.
It was put initially in meta-smartphone, but the layers configuration
doesn't actually allow this.

Bring back this bbappend in meta-webos-ports and keep it separate, so
that we can devise later if there is a better place to put this.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>