webruntime: fix app_shell close request
When Close is called on the Wayland app_shell (i.e. browser) window,
it calls the PlatformWindowDelegate::OnWindowHostClose method.

However, by default, this method isn't implemented in Aura's platform
window code, because closing a window doesn't always mean closing the app.

As this default code is only called in the browser's context, and not in
WAM, implement it to close the whole app.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>