maliit-framework-webos: improve maliit service
Restore a fix made some month ago to fix maliit's restart,
which happens when upgrading LSM package for instance.
See 28ee326 for the main fix.

Fix two things:

* maliit-server.path triggers maliit-server.service when lsm-ready
file is present. However, as LSM depends on maliit-server, it
means we'll never be able to stop the surface-manager service.
In addition, maliit-server's startup is already triggered by
surface-manager.service, so it's useless.
* having maliit-server@.service in SYSTEMD_SERVICE will insert it
in the postrm scripts of the package. However, this is a
template service, so it's incorrect.
Similarly, starting it at startup is incorrect as the instance
argument is mandatory.
So just remove it, maliit-server@0.service will be started by
maliit-server.service anyway.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>