luneos-package: pack fastboot image with lk2nd if present
The lk2nd Android bootloader is capable of chainloading an Android
kernel image; to do so, it looks for an image in the same
partition but at the offset 1Mio.

Therefore, if lk2nd is detected in the deploy folder, pack lk2nd
and the android kernel together, with a 1Mio offset.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>