luna-surfacemanager: fix startup when DRI card appears late
On rosy device, the firmware for the GPU can be loaded quite late,
and if surface-manager is started before /dev/dri/card0 exists then it
will crash. As a consequence, webapp-mgr.service will be deactivated
because its dependency has failed.

To fix that, wait that /dev/dri/card0 appears before launching
surface-manager.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>